rework imagfe endpoints as non-ninja views
