+++ core/api/peoplefinder/profile.py
@@ -1,7 +1,5 @@
 from django import forms
 from django.core.files.base import File
-from django.http import HttpResponse, HttpResponseBadRequest, HttpResponseNotFound, HttpResponseNotAllowed
-from django.core.exceptions import ValidationError
 from django.views.decorators.csrf import csrf_exempt
 from storages.backends.s3 import S3File
 from ninja import File, Router, Schema, Form
@@ -341,64 +339,6 @@ def update_profile(
 #         422: Error,
 #     },
 # )
-
-class PhotoForm(forms.Form):
-    image = forms.ImageField()
-
-    def clean(self):
-        cleaned_data = super().clean()
-        self.validate_photo(cleaned_data["image"])
-        return cleaned_data
-
-    def validate_photo(self, image):
-        if not hasattr(image, "image"):
-            return
-
-        if image.image.width < 500:
-            self.add_error("image", ValidationError("Width is less than 500px"))
-
-        if image.image.height < 500:
-            self.add_error("image", ValidationError("Height is less than 500px"))
-
-        # 8mb in bytes
-        if image.size > 1024 * 1024 * 8:
-            self.add_error("image", ValidationError("File size is greater than 8MB"))
-
-@csrf_exempt
-def upload_profile_photo(request, slug: str):
-    """
-    Endpoint to upload a new profile photo for the given profile
-    """
-    print("CALLED")
-    if request.method != "POST":
-        return HttpResponseNotAllowed(["POST", "DELETE"])
-
-    try:
-        profile = core_services.get_peoplefinder_profile_by_slug(slug=slug)
-    except PeopleFinderProfile.DoesNotExist:
-        return HttpResponseNotFound(json.dumps({
-            "message": "Unable to find people finder profile",
-        }))
-
-    form = PhotoForm(request.POST, request.FILES)
-    if not form.is_valid():
-        return HttpResponseBadRequest(json.dumps(form.errors))
-
-    # image = request.FILES["file"]
-
-    # try:
-    #     im = Image.open(image)
-    #     im.verify()
-    # except:
-    #     return 422, {
-    #         "message": "Not a valid image file",
-    #     }
-
-    # photo: File = image.open()
-    profile.photo.delete()
-    profile.photo = form.cleaned_data["image"]
-    profile.save()
-    return HttpResponse(ProfileResponse.from_orm(profile).json())
 
 
 # @profile_router.delete(
